potential fix for vite emulation
