+++ React/src/App.jsx
@@ -6,29 +6,18 @@ import './App.css'
 import { account } from './Components/appwrite-oauth/appwrite-config.js';
 import UserInfo from './Components/appwrite-oauth/appwrite-session-info.jsx';
 
-import GetPlayback from './Pages/spotify-player.jsx';
 import Homepage from './Pages/homepage.jsx';
 import Account from './Pages/account.jsx';
 import Navbar from './Components/page-elements/navbar.jsx';
 
-export const sessionContext = createContext(undefined);
+import { SessionProvider } from './logic-necessary/session-provider.jsx';
 
 
 function App() {
-    const [session, setSession] = useState(null);
     const navigate = useNavigate(); // Use navigate to redirect
     const location = useLocation();
 
     useEffect(() => {
-        const checkSession = async () => {
-            try {
-                const currentSession = await account.getSession('current');
-                console.log('Current session:', currentSession);
-                setSession(currentSession); // Save session info
-            } catch (error) {
-                console.log('No active session:', error);
-            }
-        };
         const initializeNavbar = () => {
             let currentPath = location.pathname;
             if (currentPath == '/') {
@@ -43,7 +32,6 @@ function App() {
             document.getElementById(currentPath).className = "active";
         };
 
-        checkSession();
         initializeNavbar()
     }, []);
 
@@ -54,39 +42,21 @@ function App() {
     return (
         <div className="App">
             <Navbar />
-            <div className = "Page">
-                <Routes>
-                    <Route path="/playback" element={
-                        <>
-                            {session && <GetPlayback authToken={session.providerAccessToken}  />}
-                        </>
-                    } />
-
-                    <Route path="/account" element={
-                        <sessionContext.Provider value={{ session, setSession }}>
+            <div className="Page">
+                <SessionProvider>
+                    <Routes>
+                        <Route path="/account" element={
                             <Account />
-                        </sessionContext.Provider>
-                    } />
 
-                    <Route path="/success" element={
-                        <>
-                            <div>Login Successful</div>
-                            {session && <UserInfo/>} {/* Render UserInfo here */}
-                            <button onClick={goToPlayback}>Go To PlayBack Data</button>
-                        </>
-                    } />
+                        } />
 
-                    <Route path="/failed" element={<div>Login Failed</div>} />
+                        <Route path="/failed" element={<div>Login Failed</div>} />
 
-                    <Route path="/" element={
-                        <>
-                            <sessionContext.Provider value={{session, setSession}}>
-                                <Homepage />
-                            </sessionContext.Provider>
-                        
-                        </>
-                    } />                
-                </Routes>
+                        <Route path="/" element={
+                            <Homepage />
+                        } />                
+                    </Routes>
+                </SessionProvider>
             </div>
         </div>
     );
