+++ React/src/Pages/account.jsx
@@ -5,10 +5,10 @@ import OAuthLogin from '../Components/appwrite-oauth/appwrite-spotify.jsx';
 import OAuthLogOut from '../Components/appwrite-oauth/appwrite-logout.jsx';
 import UserInfo from '../Components/appwrite-oauth/appwrite-session-info.jsx';
 
-import { sessionContext } from '../App.jsx';
+import { UseSessionContext } from '../logic-necessary/session-provider.jsx';
 
 const Account = () => {
-    const sessionInfo = useContext(sessionContext);
+    const sessionInfo = UseSessionContext();
     const navigate = useNavigate();
 
     const handleLoginSuccess = (currentSession) => {
@@ -21,15 +21,26 @@ const Account = () => {
         navigate('/account'); // Redirect to account route
     };
 
-    return (
-        <>
-            {!sessionInfo.session && <h3>Not Logged In!</h3>}
-            {!sessionInfo.session && <OAuthLogin onLoginSuccess={handleLoginSuccess} />}
-            {sessionInfo.session && <h3>Logged In!</h3>}
-            {sessionInfo.session && <UserInfo />}
-            {sessionInfo.session && <OAuthLogOut onLogOutSuccess={handleLogOutSuccess} />}
-        </>
-    );
+    if (sessionInfo) {
+        return (
+            <>
+                {!sessionInfo.session && <h3>Not Logged In!</h3>}
+                {!sessionInfo.session && <OAuthLogin onLoginSuccess={handleLoginSuccess} />}
+                {sessionInfo.session && <h3>Logged In!</h3>}
+                {sessionInfo.session && <UserInfo />}
+                {sessionInfo.session && <OAuthLogOut onLogOutSuccess={handleLogOutSuccess} />}
+            </>
+        );
+    }
+    else {
+        return (
+            <>
+                <h3>Not Logged In!</h3>
+                <OAuthLogin onLoginSuccess={handleLoginSuccess} />
+            </>
+        )
+    }
+    
 };
 
 export default Account
